+++ src/lib/general/args/map_foreach.c
@@ -24,12 +24,12 @@ t_err	minirt_args_map_foreach(
 	size_t	i;
 	size_t	j;
 
-	i = -1;
+	i = (size_t)-1;
 	while (++i < self->map_count)
 	{
 		if (minirt_str_eq(self->map[i].option->name, name))
 		{
-			j = -1;
+			j = (size_t)-1;
 			while (++j < self->map[i].entry_count)
 			{
 				if (func(context, self->map[i].entries[j].key,
