+++ src/lib/general/args/map_get.c
@@ -23,12 +23,12 @@ const char	*minirt_args_map_get(
 	size_t	i;
 	size_t	j;
 
-	i = -1;
+	i = (size_t)-1;
 	while (++i < self->map_count)
 	{
 		if (minirt_str_eq(self->map[i].option->name, name))
 		{
-			j = -1;
+			j = (size_t)-1;
 			while (++j < self->map[i].entry_count)
 				if (minirt_str_eq(self->map[i].entries[j].key, key))
 					return (self->map[i].entries[j].value);
