+++ src/lib/general/args/set_has.c
@@ -23,12 +23,12 @@ bool	minirt_args_set_has(
 	size_t	i;
 	size_t	j;
 
-	i = -1;
+	i = (size_t)-1;
 	while (++i < self->set_count)
 	{
 		if (minirt_str_eq(self->set[i].option->name, name))
 		{
-			j = -1;
+			j = (size_t)-1;
 			while (++j < self->set[i].value_count)
 				if (minirt_str_eq(self->set[i].values[j], value))
 					return (true);
